requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1
astor==0.8.1
black==23.12.1
pydantic==2.5.0
rich==13.7.0
typer==0.9.0
pyyaml==6.0.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
